refactor: queryインターフェースの配置変更
refs #008
