+++ MobileSuitDatabase/src/main/java/jp/co/sunarch/mobilesuitDatabase/port/adapter/api/controller/internal/ArmsCountController.java
@@ -1,0 +1,9 @@
+package jp.co.sunarch.mobilesuitDatabase.port.adapter.api.controller.internal;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class ArmsCountController {
+
+	// 武器登録件数取得処理を呼び出す
+}
