+++ MobileSuitDatabase/src/main/java/jp/co/sunarch/mobilesuitDatabase/port/adapter/query/arms/ArmsQuery.java
@@ -1,4 +1,4 @@
-package jp.co.sunarch.mobilesuitDatabase.port.adapter.web.controller.arms;
+package jp.co.sunarch.mobilesuitDatabase.port.adapter.query.arms;
 
 import java.util.List;
