+++ MobileSuitDatabase/src/main/java/jp/co/sunarch/mobilesuitDatabase/port/adapter/web/controller/arms/DeleteArmsController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import jp.co.sunarch.mobilesuitDatabase.application.command.arms.DeleteArmsCommand;
 import jp.co.sunarch.mobilesuitDatabase.application.usecase.arms.DeleteArmsUseCase;
 import jp.co.sunarch.mobilesuitDatabase.domain.model.arms.ArmsId;
+import jp.co.sunarch.mobilesuitDatabase.port.adapter.query.arms.ArmsQuery;
 import jp.co.sunarch.mobilesuitDatabase.port.adapter.web.model.arms.ArmsModel;
 import lombok.RequiredArgsConstructor;
