+++ MobileSuitDatabase/src/test/java/jp/co/sunarch/mobilesuitDatabase/port/adapter/web/controller/arms/DeleteArmsControllerTest.java
@@ -18,6 +18,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import jp.co.sunarch.mobilesuitDatabase.application.command.arms.DeleteArmsCommand;
 import jp.co.sunarch.mobilesuitDatabase.application.usecase.arms.DeleteArmsUseCase;
 import jp.co.sunarch.mobilesuitDatabase.domain.model.arms.ArmsId;
+import jp.co.sunarch.mobilesuitDatabase.port.adapter.query.arms.ArmsQuery;
 import jp.co.sunarch.mobilesuitDatabase.port.adapter.web.model.arms.ArmsModel;
 
 @WebMvcTest(DeleteArmsController.class)
